+++ src/main/java/selfConstructed/SkySalerADS/service/CommentService.java
@@ -3,9 +3,9 @@ package selfConstructed.SkySalerADS.service;
 import selfConstructed.SkySalerADS.dto.CommentDTO;
 
 public interface CommentService {
-    CommentDTO createComment(CommentDTO commentDTO, Integer adId, Integer userId);
+    CommentDTO createComment(CommentDTO commentDTO, Long adId, Long userId);
 
     CommentDTO updateComment(CommentDTO commentDTO);
 
-    void deleteComment(Integer commentId, Integer userId);
+    void deleteComment(Long commentId, Long userId);
 }
